requirements.txt
streamlit
pandas
firebase-admin
python-dotenv
requests
apscheduler
streamlit-google-auth
